+++ pages/index.vue
@@ -8,6 +8,9 @@
 </template>
 
 <script setup lang="ts">
+import { CommunityEvent } from "../types/events";
+import rawUpcomingEvents from "~/content/events/upcoming-community-events.json";
+
 interface PackageInfo {
   info: {
     version: string;
@@ -39,6 +42,56 @@ const { data: packageInfo } = await useAsyncData<PackageInfo>(() =>
 
 const qiskitVersion = packageInfo.value?.info.version ?? "";
 
+const upcomingEvents = rawUpcomingEvents as CommunityEvent[];
+
+const createEventSchema = (events: CommunityEvent[]) => {
+  const entities = events
+    .filter((event) => event.startDate)
+    .map((event) => {
+      let location;
+      if (["YouTube", "Virtual"].includes(event.location)) {
+        location = defineVirtualLocation({
+          url: event.to,
+        });
+      } else {
+        location = definePlace({
+          ame: event.location,
+          address: event.location,
+          url: event.to,
+        });
+      }
+
+      const eventAttendanceMode = ["YouTube", "Virtual"].includes(
+        event.location
+      )
+        ? "OnlineEventAttendanceMode"
+        : "OfflineEventAttendanceMode";
+
+      const schemaEvent: any = {
+        name: event.title,
+        endDate: new Date(event.endDate),
+        eventAttendanceMode,
+        image: event.image,
+        location,
+        startDate: new Date(event.startDate),
+        organizer: {
+          name: "IBM Quantum",
+          url: "https://ibm.com/quantum",
+        },
+      };
+
+      if (event.speaker) {
+        schemaEvent.performer = event.speaker;
+      }
+
+      return defineEvent(schemaEvent);
+    });
+
+  return entities;
+};
+
+const eventsSchema = createEventSchema(upcomingEvents);
+
 useSchemaOrg([
   defineWebSite({
     name: "Qiskit.org",
@@ -48,5 +101,6 @@ useSchemaOrg([
     logo: "/images/qiskit-logo.png",
   }),
   defineWebPage(),
+  ...eventsSchema,
 ]);
 </script>
